♻️ Refactor(): add treasure management to character schema and seeder
