+++ services/adventure/api/src/resources/character/core/schemas/character.schema.ts
@@ -9,7 +9,7 @@ import { Stats } from '@/resources/character/core/schemas/stats/stats.schema';
 import { ApiProperty } from '@nestjs/swagger';
 import { Appearance } from '@/resources/character/player/schemas/appearance/appearance.schema';
 import { Background } from '@/resources/character/player/schemas/background/background.schema';
-import { Treasure } from '@/resources/character/player/schemas/treasure/treasure.schema';
+import { Treasure } from '@/resources/character/core/schemas/treasure/treasure.schema';
 import { Conditions } from '@/resources/character/core/schemas/conditions/conditions.schema';
 
 export type CharacterDocument = Character & Document;
